backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
celery==5.3.4
redis==5.0.1
scrapy==2.11.0
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
fastapi-cors==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
